Publish XLSForm

Publish XLSForm

ona_form_publish(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  xls_file = NULL,
  xls_url = NULL,
  dropbox_xls_url = NULL,
  project_id = NULL,
  public = FALSE
)

Arguments

base_url

ONA URL. Default is https://api.ona.io.

auth_mode

Password or token? Default is token.

xls_file

Path to the XLSForm file.

xls_url

URL to the XLSForm file.

dropbox_xls_url

Dropbox URL to the XLSForm file.

project_id

Project identifier of project to publish XLSForm to. If NULL (default), XLSForm is published to the an account's default project.

public

Logical. Should the form be public? Default to FALSE.

Value

A published form on ONA.

Examples

project_list <- ona_project_list()
if ("test" %in% project_list$name) {
  project_id = project_list$id[project_list$name == "test"]
  ona_project_delete(project_id = project_id)
}
#> Warning: Unknown or uninitialised column: `id`.
#> Response [https://api.ona.io/api/v1/projects/]
#>   Date: 2023-10-21 06:49
#>   Status: 404
#>   Content-Type: text/html; charset=utf-8
#>   Size: 4.58 kB
#> <!DOCTYPE html>
#> <html lang="en">
#>   <head>
#>     <meta charset="utf-8">
#>     <title>api.ona.io</title>
#>     <meta name="description" content="Mobile data collection made easy. formh...
#>     <meta name="author" content="">
#>     
#>     <meta name="google-site-verification" content="H0JRFrqLHnVfHuZaHRfz1Gmc5f...
#>     
#> ...
ona_project_register(name = "test")
#> Warning: Specified project name already in use in the ernestguevarra ONA account. Try a different project name.
#> $non_field_errors
#> [1] "The fields name, owner must make a unique set."
#> 
project_list <- ona_project_list()
project_id = project_list$id[project_list$name == "test"]
#> Warning: Unknown or uninitialised column: `id`.
ona_form_publish(
  xls_file = system.file(
    "appearance_widgets.xlsx", package = "okapi"
  ),
  project_id = project_id
)
#> $url
#> [1] "https://api.ona.io/api/v1/forms/771091"
#> 
#> $formid
#> [1] 771091
#> 
#> $metadata
#> list()
#> 
#> $owner
#> [1] "https://api.ona.io/api/v1/users/ernestguevarra"
#> 
#> $created_by
#> [1] "https://api.ona.io/api/v1/users/ernestguevarra"
#> 
#> $public
#> [1] FALSE
#> 
#> $public_data
#> [1] FALSE
#> 
#> $public_key
#> [1] ""
#> 
#> $require_auth
#> [1] FALSE
#> 
#> $submission_count_for_today
#> [1] 0
#> 
#> $tags
#> list()
#> 
#> $title
#> [1] "Appearance Widgets"
#> 
#> $users
#>   is_org metadata.zebra.tos-pp.nil.date metadata.zebra.tos-pp.nil.accepted
#> 1  FALSE       2021-01-07T09:43:19.973Z                               TRUE
#>   metadata.zebra.tos-pp.2018-03-30.date
#> 1              2020-09-17T11:52:25.034Z
#>   metadata.zebra.tos-pp.2018-03-30.accepted
#> 1                                      TRUE
#>   metadata.zebra.eu-citizen-consent.date
#> 1               2020-09-17T11:52:23.960Z
#>   metadata.zebra.eu-citizen-consent.eu-citizen-data
#> 1                                             FALSE
#>   metadata.zebra.hide-home-walkthrough?
#> 1                                  TRUE
#>   metadata.zebra.hide-data-view-walkthrough?
#> 1                                       TRUE
#>   metadata.zebra.hide-share-modal-walkthrough?
#> 1                                         TRUE
#>   metadata.zebra.hide-project-view-walkthrough?
#> 1                                          TRUE
#>   metadata.zebra.hide-beta-features-walkthrough?
#> 1                                           TRUE
#>   metadata.zebra.hide-dataview-beta-tag-walkthrough?
#> 1                                               TRUE
#>                                                                 metadata.avatar-url
#> 1 https://images.ona.io/image/a9af299945734cfb9d5869bc7c3eaca8/Guevarra Photo 2.jpg
#>   metadata.is_email_verified      metadata.last_password_edit first_name
#> 1                       TRUE 2022-03-28T11:23:57.241759+00:00     Ernest
#>   last_name           user  role
#> 1  Guevarra ernestguevarra owner
#> 
#> $enketo_url
#> [1] "https://enketo.ona.io/x/HApYyTFd"
#> 
#> $enketo_preview_url
#> [1] "https://enketo.ona.io/preview/HApYyTFd"
#> 
#> $enketo_single_submit_url
#> [1] "https://enketo.ona.io/x/HApYyTFd"
#> 
#> $num_of_submissions
#> [1] 0
#> 
#> $last_submission_time
#> NULL
#> 
#> $form_versions
#> list()
#> 
#> $data_views
#> list()
#> 
#> $xls_available
#> [1] TRUE
#> 
#> $has_id_string_changed
#> [1] FALSE
#> 
#> $description
#> [1] ""
#> 
#> $downloadable
#> [1] TRUE
#> 
#> $allows_sms
#> [1] FALSE
#> 
#> $encrypted
#> [1] FALSE
#> 
#> $sms_id_string
#> [1] "appearance_widgets"
#> 
#> $id_string
#> [1] "appearance_widgets"
#> 
#> $date_created
#> [1] "2023-10-21T06:49:25.987613Z"
#> 
#> $date_modified
#> [1] "2023-10-21T06:49:25.987634Z"
#> 
#> $uuid
#> [1] "d8495ff3fd284763bce738cc6d3a2ff6"
#> 
#> $bamboo_dataset
#> [1] ""
#> 
#> $instances_with_geopoints
#> [1] FALSE
#> 
#> $instances_with_osm
#> [1] FALSE
#> 
#> $version
#> [1] "202310210649"
#> 
#> $has_hxl_support
#> [1] FALSE
#> 
#> $last_updated_at
#> [1] "2023-10-21T06:49:25.987680Z"
#> 
#> $hash
#> [1] "md5:1e560cf0b203b0bd010a23aa5feca20a"
#> 
#> $is_merged_dataset
#> [1] FALSE
#> 
#> $is_instance_json_regenerated
#> [1] FALSE
#> 
#> $project
#> [1] "https://api.ona.io/api/v1/projects/220292"
#>